refactor into PlaylistCard
